+++ emukit/core/initial_designs/__init__.py
@@ -1,0 +1,5 @@
+# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+from .latin_design import LatinDesign  # noqa: F401
+from .random_design import RandomDesign  # noqa: F401
