+++ emukit/examples/fabolas/fmin.py
@@ -3,7 +3,7 @@ import numpy as np
 from emukit.core.loop import FixedIterationsStoppingCondition, UserFunctionWrapper, LoopState
 from emukit.core.parameter_space import ParameterSpace
 from emukit.examples.fabolas import FabolasLoop
-from emukit.experimental_design.model_free.latin_design import LatinDesign
+from emukit.core.initial_designs import LatinDesign
 
 
 def fmin_fabolas(func, space: ParameterSpace, s_min: float, s_max: float, n_iters: int,
