+++ emukit/experimental_design/acquisitions/__init__.py
@@ -1,0 +1,6 @@
+# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+
+from .model_variance import ModelVariance  # noqa: F401
+from .integrated_variance import IntegratedVarianceReduction  # noqa: F401
