--- emukit/experimental_design/model_based/__init__.py
@@ -1,5 +1,0 @@
-# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-
-from .experimental_design_loop import ExperimentalDesignLoop  # noqa: F401
