--- emukit/experimental_design/model_free/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
