+++ tests/emukit/experimental_design/test_experimental_design_loop.py
@@ -2,8 +2,8 @@ import GPy
 import numpy as np
 from numpy.testing import assert_array_equal
 
-from emukit.experimental_design.model_based.acquisitions import ModelVariance
-from emukit.experimental_design.model_based import ExperimentalDesignLoop
+from emukit.experimental_design.acquisitions import ModelVariance
+from emukit.experimental_design import ExperimentalDesignLoop
 from emukit.core.parameter_space import ParameterSpace
 from emukit.core.continuous_parameter import ContinuousParameter
 from emukit.core.loop import UserFunctionWrapper, FixedIterationsStoppingCondition
