+++ tests/emukit/experimental_design/test_integrated_variance.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from emukit.experimental_design.model_based.acquisitions.integrated_variance import IntegratedVarianceReduction
+from emukit.experimental_design.acquisitions.integrated_variance import IntegratedVarianceReduction
 
 
 def test_integrated_variance_fails_with_out_of_domain_test_points(gpy_model, continuous_space):
